+++ src/main/scala/carldata/sf/face/FaceParser.scala
@@ -15,8 +15,8 @@ object FaceParser extends RegexParsers {
   }
 
   def funParams: Parser[Seq[Expression]] = repsep(addOrBoolExpr, ",")
-  def number: Parser[NumberLiteral] = """[+-]?([0-9]*[.])?[0-9]+""".r ^^ { ds =>  NumberLiteral(ds.toFloat) }
-  def nullVariable: Parser[NumberLiteral] = ("null" | "NULL") ^^ { _ => NumberLiteral(Float.NaN)}
+  def number: Parser[NumberLiteral] = """[+-]?([0-9]*[.])?[0-9]+""".r ^^ { ds =>  NumberLiteral(ds.toDouble) }
+  def nullVariable: Parser[NumberLiteral] = ("null" | "NULL") ^^ { _ => NumberLiteral(Double.NaN)}
   def variable: Parser[VariableExpr] = identifier ^^ { id => VariableExpr(id) }
   def negativeExpr: Parser[MinusOpExpr] = "-" ~ factor ^^ { id => MinusOpExpr(id._2) }
   def factor: Parser[Expression] = function | number | nullVariable | variable | negativeExpr | "(" ~> addOrBoolExpr <~ ")"
