+++ src/test/scala/carldata/sf/face/FaceParserTest.scala
@@ -124,21 +124,21 @@ class FaceParserTest extends FlatSpec with Matchers {
   it should "parse expression with null" in {
     val code = "if(A == NULL ,0,A)"
     val result = FaceParser.parse(code)
-    val expected = AppExpr("if", Seq(RelationExpr(VariableExpr("A"), "==", NumberLiteral(Float.NaN)), NumberLiteral(0), VariableExpr("A")))
+    val expected = AppExpr("if", Seq(RelationExpr(VariableExpr("A"), "==", NumberLiteral(Double.NaN)), NumberLiteral(0), VariableExpr("A")))
     result.toString shouldBe Right(expected).toString
   }
 
   it should "parse expression with number in .x format" in {
     val code = "if(A == .05 ,0,A)"
     val result = FaceParser.parse(code)
-    val expected = AppExpr("if", Seq(RelationExpr(VariableExpr("A"), "==", NumberLiteral(0.05f)), NumberLiteral(0), VariableExpr("A")))
+    val expected = AppExpr("if", Seq(RelationExpr(VariableExpr("A"), "==", NumberLiteral(0.05)), NumberLiteral(0), VariableExpr("A")))
     result.toString shouldBe Right(expected).toString
   }
 
   it should "parse expression with negative number in .x format" in {
     val code = "if(A == -.05 ,0,A)"
     val result = FaceParser.parse(code)
-    val expected = AppExpr("if", Seq(RelationExpr(VariableExpr("A"), "==", NumberLiteral(-0.05f)), NumberLiteral(0), VariableExpr("A")))
+    val expected = AppExpr("if", Seq(RelationExpr(VariableExpr("A"), "==", NumberLiteral(-0.05)), NumberLiteral(0), VariableExpr("A")))
     result.toString shouldBe Right(expected).toString
   }
